add getChildren static method for VictoryGroup
